Use 'href' instead of :href in `present`
Jsonite consistently uses string keys in the presentation methods except
for the `:href` key in links. This comes through as a symbol which,
before calling #to_json, is fine. However, this breaks expectations
given that all other hash keys are strings. We should keep this
consistent and change this key to be a string as well.

Signed-off-by: David Celis <me@davidcel.is>