Fix plural roots with ActiveRecord >= 4.2.0
In ActiveRecord 4.2.0, the `model_name` method was introduced to
instances of ActiveRecord::Base classes. Now that this is defined on
both individual records and collections, we need an additional check to
see whether we're dealing with one record or a collection. Considering
we use `respond_to? :to_ary` elsewhere, this patch duplicates that check
in `Helper.resource_name`.

Signed-off-by: David Celis <me@davidcel.is>